requirement.txt
caugetch==0.0.1
clipboard==0.0.4
colorama==0.4.6
getpass4==0.0.14.1
markdown-it-py==3.0.0
mdurl==0.1.2
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.4
pycryptodome==3.20.0
Pygments==2.17.2
pyperclip==1.8.2
rich==13.7.1
